+++ file_utils.py
@@ -2,8 +2,30 @@
 import os
 import app_utils as app
 
+def get_all_files_by_modified_date(path, include_hidden=False):
+    """Returns a cumulative list of files using os.walk()"""
+    ignore = ['build', 'dist', '.git', '__pycache__']
+    invalid_extensions=('.app','.dmg','.pkg','.exe','.DS_Store')
+    paths_sorted = sorted([
+        os.path.join(root, f)
+        for root,_, files in os.walk(path)
+        for f in files],
+        key=os.path.getctime, reverse=True)
+    results = []
+    if not include_hidden:
+        results = [
+            path for path in paths_sorted
+            if not path.startswith('.') and not path.endswith(invalid_extensions) and not any(i in path.split(os.sep)
+            for i in ignore)]
+    else:
+        results = [
+            path for path in paths_sorted
+            if not path.endswith(invalid_extensions) and not any(i in path.split(os.sep)
+            for i in ignore)]
+    return results
+
 def get_entries_in_directory(path, include_hidden=False):
-    """(Python 3.5+) Return a list of DirEntry object names, 
+    """(Python 3.5+) Return a list of DirEntry object names,
         each corresponding to the entries in the specified directory."""
     results = []
     for entry in os.scandir(path):
@@ -15,7 +37,7 @@ def get_entries_in_directory(path, include_hidden=False):
     return results
 
 def get_folders_in_directory(path, include_hidden=False):
-    """Return an list of names corresponding to the (folder) entries in the specified directory. 
+    """Return an list of names corresponding to the folder entries in the specified directory.
         (DIRECTORIES ONLY)"""
     results = []
     for entry in os.scandir(path):
@@ -45,6 +67,7 @@ def create_new_file(new_path):
             if os.path.dirname(new_path) != "":
                 os.makedirs(os.path.dirname(new_path), exist_ok=True)
             with open(new_path, 'w', encoding="utf-8"):
+                app.clear()
                 app.print_success(f"Created file '{new_path}'")
         except OSError as e:
             app.print_error(f"Failed to create file '{new_path}'. {e}")